+++ src/seedsigner/gui/screens/settings_screens.py
@@ -364,28 +364,6 @@ class IOTestScreen(BaseTopNavScreen):
 
 
 @dataclass
-class DonateScreen(BaseTopNavScreen):
-    def __post_init__(self):
-        self.title = _("Donate")
-        super().__post_init__()
-
-        self.components.append(TextArea(
-            # TRANSLATOR_NOTE: If your language uses the percent sign ("%"), your translation must also use two percent signs ("%%") due to python formatting oddities. "100%%" will be rendered as "100%".
-            text=_("SeedSigner is 100%% free & open source, funded solely by the Bitcoin community.\n\nDonate onchain or LN at:").replace("%%", "%"),
-            screen_y=self.top_nav.height + 3*GUIConstants.COMPONENT_PADDING,
-        ))
-
-        self.components.append(TextArea(
-            text="seedsigner.com",
-            font_name=GUIConstants.get_body_font_name(),
-            font_size=28,
-            font_color=GUIConstants.ACCENT_COLOR,
-            supersampling_factor=1,
-            screen_y=self.components[-1].screen_y + self.components[-1].height + GUIConstants.COMPONENT_PADDING
-        ))
-
-
-@dataclass
 class BatteryInfoScreen(BaseTopNavScreen):
     """Display live battery information from the UPS HAT."""