+++ tests/test_flows_settings.py
@@ -38,24 +38,6 @@ class TestSettingsFlows(FlowTest):
         assert os.path.exists(Settings.SETTINGS_FILENAME) == True
 
 
-    def test_multiselect(self):
-        """ Multiselect Settings options should stay in-place; requires BACK to exit. """
-        # Which option are we testing?
-        settings_entry = SettingsDefinition.get_settings_entry(SettingsConstants.SETTING__SEED_WORD_LENGTHS)
-
-        self.run_sequence([
-            FlowStep(MainMenuView, button_data_selection=MainMenuView.SETTINGS),
-            # Seed word lengths lives under the Advanced submenu.
-            FlowStep(settings_views.SettingsMenuView, button_data_selection=settings_views.SettingsMenuView.ADVANCED),
-            FlowStep(settings_views.SettingsMenuView, button_data_selection=ButtonOption(settings_entry.display_name)),
-            FlowStep(settings_views.SettingsEntryUpdateSelectionView, screen_return_value=0),  # select/deselect first option
-            FlowStep(settings_views.SettingsEntryUpdateSelectionView, screen_return_value=1),  # select/deselect second option
-            FlowStep(settings_views.SettingsEntryUpdateSelectionView, screen_return_value=1),  # select/deselect second option
-            FlowStep(settings_views.SettingsEntryUpdateSelectionView, screen_return_value=RET_CODE__BACK_BUTTON),  # BACK to exit
-            FlowStep(settings_views.SettingsMenuView),
-        ])
-
-
     def test_io_test(self):
         """ Basic flow from MainMenuView to I/O Test View """
         self.run_sequence([
@@ -102,37 +84,6 @@ class TestSettingsFlows(FlowTest):
             FlowStep(settings_views.SettingsMenuView, screen_return_value=RET_CODE__BACK_BUTTON),
             FlowStep(settings_views.SettingsMenuView, before_run=assert_general),
         ])
-
-
-    def test_donate(self):
-        """ Basic flow from MainMenuView to Donate View """        
-        self.run_sequence([
-            FlowStep(MainMenuView, button_data_selection=MainMenuView.SETTINGS),
-            FlowStep(settings_views.SettingsMenuView, button_data_selection=settings_views.SettingsMenuView.DONATE),
-            FlowStep(settings_views.DonateView),
-            FlowStep(settings_views.SettingsMenuView),
-        ])
-
-
-    def test_load_backup_files_submenu(self):
-        tapsigner_entry = SettingsDefinition.get_settings_entry(SettingsConstants.SETTING__TAPSIGNER_BACKUP)
-
-        self.run_sequence([
-            FlowStep(MainMenuView, button_data_selection=MainMenuView.SETTINGS),
-            FlowStep(settings_views.SettingsMenuView, button_data_selection=settings_views.SettingsMenuView.ADVANCED),
-            FlowStep(settings_views.SettingsMenuView, button_data_selection=settings_views.SettingsMenuView.LOAD_BACKUP_FILES),
-            FlowStep(settings_views.LoadBackupFilesSettingsView, button_data_selection=ButtonOption(tapsigner_entry.display_name)),
-            FlowStep(settings_views.SettingsEntryUpdateSelectionView, button_data_selection=ButtonOption(tapsigner_entry.get_selection_option_display_name_by_value(SettingsConstants.OPTION__ENABLED))),
-            FlowStep(settings_views.SettingsEntryUpdateSelectionView, screen_return_value=RET_CODE__BACK_BUTTON),
-            FlowStep(settings_views.LoadBackupFilesSettingsView, screen_return_value=RET_CODE__BACK_BUTTON),
-            FlowStep(settings_views.SettingsMenuView),
-        ])
-
-        assert self.settings.get_value(SettingsConstants.SETTING__TAPSIGNER_BACKUP) == SettingsConstants.OPTION__ENABLED
-
-
-    def test_tapsigner_backup_setting_default_disabled(self):
-        assert self.settings.get_value(SettingsConstants.SETTING__TAPSIGNER_BACKUP) == SettingsConstants.OPTION__DISABLED
 
 
     def test_settingsqr(self):